[AIRLFOW-YYY] Move worker command to separate file
